% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUnitMeans.R
\name{addUnitMeans}
\alias{addUnitMeans}
\title{Add cluster/group/person means to a data frame for a set of specified variables}
\usage{
addUnitMeans(data, unitVariable, meanVariables, newNames)
}
\arguments{
\item{data}{A data frame containing numeric variables and a unit ID variable}

\item{unitVariable}{A variable containing the ID number for each cluster/group/person}

\item{meanVariables}{A character vector containing the names of the variables for which
the means for each cluster/group/person are needed}

\item{newNames}{A character vector of the same length as \code{meanVariables} with the names
for the new variables of cluster/group/person means}
}
\value{
A data frame where new variables are added
}
\description{
Add cluster/group/person means to a data frame
}
\examples{

# load data for example 3a
data(example3aData)

# display variable names for data set
names(example3aData)

# add school means with this function
analysisData = addUnitMeans(
  data=studentsZ,
  unitVariable = "schoolID",
  meanVariables = c("langpost", "IQverbz", "IQperfz"),
  newNames = c("SMlangpost", "SMIQverbz", "SMIQperfz")
)

# display new names
names(analysisData)

}
