% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoRSquaredInator.R
\name{pseudoRSquaredinator}
\alias{pseudoRSquaredinator}
\title{Provides pseudo R squared for comparing two models (one nested within the other)}
\usage{
pseudoRSquaredinator(smallerModel, largerModel)
}
\arguments{
\item{smallerModel}{A model (from \code{lme4::lmer}) that has a subset of predictors from \code{largerModel}}

\item{largerModel}{A  model (from \code{lme4::lmer}) that has all predictors from \code{smallerModel} and additional predictors}
}
\value{
A list with all calculated values along with a display of these values to the console window
}
\description{
Provides pseudo R squared for comparing two models (one nested within the other).
}
\examples{

# load data for example 3a
data(example3aData)

# remove all NA values from langpost and
newData = example3aData[complete.cases(example3aData[,c("langpost", "homework")]),]

# estimate empty model with random intercept for school
emptyModel = lmer(formula = langpost ~ 1 + (1|schoolID), data = newData, REML = TRUE)
summary(emptyModel)

# add school-level homework variable as predictor
hwModel = lmer(formula = langpost ~ 1 + homework + (1|schoolID), data = newData, REML = TRUE)
summary(hwModel)

# calculate pseudo R squared using this function
hwVSempty = pseudoRSquaredinator(smallerModel = emptyModel, largerModel = hwModel)

# OUTPUT
# Pseudo R2 Estimates
# R2 Random.(Intercept): 0.00179
# R2 L1.sigma2: 5e-05

}
