% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomInterceptor.R
\name{randomInterceptor}
\alias{randomInterceptor}
\title{Estimate empty models with and without random intercepts}
\usage{
randomInterceptor(data, outcome, level2ID, REML = TRUE)
}
\arguments{
\item{data}{A data frame containing numeric variables and a unit ID variable}

\item{outcome}{A character variable containing the name of the outcome variable from \code{data}}

\item{level2ID}{A character variable containing the name of the level-two ID variable from \code{data}}

\item{REML}{A logical value where \code{TRUE} (default) estimates the models using REML and
\code{FALSE} estimates the models using ML}
}
\value{
A list with all calculated values along with a display of these values to the console window
}
\description{
Estimates empty models with and without random intercepts across packages and reports ICCs.
Calculates intraclass correlation and reports likelihood ratio test statistic, degrees of
freedom, and p-value.
}
\examples{

# load data for example 3a
data(example3aData)

# remove all NA values from langpost
newData = example3aData[complete.cases(example3aData$langpost),]

# estimate models using REML with langpost DV and schoolID as level-two ID variable:
initLangpost = randomInterceptor(data = example3aData,
                                 outcome = "langpost",
                                 level2ID = "schoolID",
                                 REML = TRUE)
# OUTPUT
# Empty Level-1 vs Empty Level-2 Random Intercept Model Comparison
# ICC: 0.2191
# Test Statistic: 502.90794
# Test df: 1
# Test p-value: 0

}
